Show E/W/I/? columns only when non-zero entries
The E/W/I/? columns are shown only when E/W/I are non-zero in at least
one quickfix list. Otherwise the column is omitted.

For example, if none of the quickfix lists contain any quickfix entries
of type 'E', the 'E' column won't be displayed. When none of the
quickfix lists contain any 'E', 'W' and 'I' entries, the '?' column is
omitted as well and only the size column is shown.