Add optional argument to override popup options
Add an optional parameter 'opts' to qfhistory#open(). The argument must
be a dictionary. Currently only the 'popup' entry is supported. It must
be a dictionary that will override some of the popup-options.

See :help popup_create-arguments for all supported popup options.

Example:

    qfhistory#open(false, {
        'popup': {
            line: &lines - &cmdheight - 5,
            pos: 'botleft',
            padding: [0, 1, 0, 1]
        }
    })

Note: Not all popup options can be passed. Entries 'filter', 'callback',
'filtermode', 'mapping' are set by the plugin and can't be changed.